Add namespace to this package
A really bad thing for a developer to do is attempt to run `npm install erc721-starter-contracts` on this package. It would be silly to do this but if it does happen they may get owned by someone else registering the package publicly. 

So lets namespace this package just to ensure a mistake like that doesn't harm the user.